+++ Pac++Man/Entity.h
@@ -8,19 +8,19 @@ Author: fookenCode
 
 class Entity {
 protected:
-    int xPos, yPos;
+    double xPos, yPos;
     bool isInvalidated;
 public:
     Entity() { }
     virtual ~Entity() { }
 
-    int getXPosition() { return xPos; }
-    int getYPosition() { return yPos; }
+    double getXPosition() { return xPos; }
+    double getYPosition() { return yPos; }
     bool IsInvalidated() { return isInvalidated; }
 
-    void setXPos(int newXPos) { xPos = newXPos; }
-    void setYPos(int newYPos) { yPos = newYPos; }
-    void setPosition(int newXPos, int newYPos) { xPos = newXPos; yPos = newYPos; }
+    void setXPos(double newXPos) { xPos = newXPos; }
+    void setYPos(double newYPos) { yPos = newYPos; }
+    void setPosition(double newXPos, double newYPos) { xPos = newXPos; yPos = newYPos; }
     void setInvalidated(bool newValue) { isInvalidated = newValue; }
 
     virtual void Render() = 0;
