+++ Pac++Man/ScoreBoard.cpp
@@ -18,8 +18,8 @@ Comments: Renders the Score information on the right-side of screen.
 void ScoreBoard::Render() {
     if (isInvalidated) {
         COORD Position;
-        Position.X = xPos;
-        Position.Y = yPos;
+        Position.X = (int)xPos;
+        Position.Y = (int)yPos;
         SetConsoleCursorPosition(GetStdHandle(STD_OUTPUT_HANDLE), Position);
         std::cout << "\033[40;37;1m" << SCORE_NAME_TEXT;
         Position.Y += 1;
